--- project/src/mocks/сities.js
@@ -1,9 +1,0 @@
-const OfferCity = {
-  PARIS: 'Paris',
-  COLOGNE: 'Cologne',
-  BRUSSELS: 'Brussels',
-  AMSTERDAM: 'Amsterdam',
-  HAMBURG: 'Hamburg',
-  DUSSELDORF: 'Dusseldorf',
-};
-export { OfferCity };
